==> 06_SBS16_sim.qmd <==
---
title: "SBS16_sim"
format: html
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
# library ######################################################################
lib_lst <- c("dplyr",
             "tibble",
             "tidyr",
             "purrr",
             "stringr",
             "ggplot2",
             "patchwork")

invisible(purrr::walk(lib_lst, library, character.only = TRUE))
rm(lib_lst)

# functions ####################################################################

# variable #####################################################################
source("config.R")
devtools::install_github("frvirard/signatR",
                         auth_token = token, 
                         upgrade = "never")
devtools::install_github("frvirard/phenotypeR", 
                         auth_token = token, 
                         upgrade = "never")
```

```{r create simulation}
project <- dir(file.path(data_dr, "/sigprofiler/matrix"))|>
  str_subset("LICA-FR_LIRI-JP_TCGA-LIHC")

dir_lst <- dir(file.path(data_dr, "sigprofiler/matrix", 
                           project,"output/simulations/"))|>
    str_subset("simulations")|>
    str_subset("384")
  
if(length(dir_lst) == 0){
  signatR::sigpro_sim(project = project,
                      project_path = file.path(data_dr, 
                                               "sigprofiler/matrix",
                                               project),
                      genome = "GRCh37",
                      contexts = list("384"), 
                      chrom_based = TRUE,
                      sim = 100)
}else{
  message("simulation folder \"", dir_lst,"\" detected, skipping ...")
}
```

```{r create annotated maf}
walk(1:100, function(run){
  phenotypeR::annotate_maf(input = paste0(run, ".maf"),
                           annot = "refGene",ref_genome = "hg19", 
                           output_dr = file.path(data_dr, "maf", "sim_tmp"),
                           input_dr = file.path(data_dr, "sigprofiler", "matrix",
                                                project, "output", 
                                                "simulations",
                                                paste0(project, "_simulations_GRCh37_384")),
                           col_select = c(Chr = "Chrom", 
                                          Start = "Start_position", 
                                          End = "End_position", 
                                          Ref = "Reference_Allele", 
                                          Alt = "Tumor_Seq_Allele2"), 
                           annovar_tool = "/Users/francoisvirard/Documents/Code/annovar")
})
```
